pyproject.toml
[project]
name = "procurement-contract-audit"
version = "0.1.0"
requires-python = ">=3.10,<3.13"
dependencies = [
    "openenv-core",
    "fastapi",
    "uvicorn",
    "pydantic>=2.0",
    "openai",
    "python-dotenv",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
    "ruff>=0.3",
    "pyright>=1.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["server"]

[tool.hatch.build.targets.wheel.force-include]
"models.py" = "models.py"
"client.py" = "client.py"
"my_env.py" = "my_env.py"

[tool.hatch.build.targets.sdist]
include = [
    "server/**",
    "models.py",
    "client.py",
    "my_env.py",
    "openenv.yaml",
    "README.md",
    "inference.py",
    "pyproject.toml",
    "Dockerfile",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
    "RUF", # ruff-specific rules
]
ignore = [
    "E501",  # line length handled by formatter
]

[tool.ruff.lint.isort]
known-first-party = ["server", "models", "client"]

[tool.pyright]
pythonVersion = "3.10"
pythonPlatform = "All"
typeCheckingMode = "basic"
include = ["server", "models.py", "client.py", "inference.py"]
exclude = ["**/__pycache__", ".venv", "tests"]
